move filters to documentation namespace
