+++ Application/EdFi.Ods.Admin.Api/Features/Connect/ConnectController.cs
@@ -3,7 +3,6 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-using EdFi.Ods.Admin.Api.ActionFilters;
 using EdFi.Ods.Admin.Api.Infrastructure.Security;
 using FluentValidation;
 using Microsoft.AspNetCore;
