+++ Application/EdFi.Ods.Admin.Api/Infrastructure/Documentation/SwaggerOptionalSchemaFilter.cs
@@ -3,11 +3,11 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
+using System.Reflection;
 using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
-using System.Reflection;
 
-namespace EdFi.Ods.Admin.Api.ActionFilters
+namespace EdFi.Ods.Admin.Api.Infrastructure.Documentation
 {
     [AttributeUsage(AttributeTargets.Property)]
     public class SwaggerOptionalAttribute : Attribute
